Update NeuroLupBatteryResults.md
My first pass adding some additional notes on the background info of some of these assessments. Perhaps in the future we will create a section specifically for this background info. I also wonder whether we should mention the exact amount of exercises included, as that means we have to update each time we add a new one.